Introduce custom IdGenerator
kotlin.uuid.Uuid doesn't work on http, only https. Also all our ID's are generated on a single thread which makes it a lot simpler.
This is a first draft. The IdGenerator should probably be part of the state so it correctly rolls back on Undo and also so we can
generate the same ID's across multiple machines