Create positive list for Ktlint
Bump Ktlint to latest version and create a positive list from their list of rules rather than a negative list. This goes against the design philosophy of Ktlint, but this project deviates from the "standard" code format in quite a few places.
This is not an accident, but happens because in some places, like in Procedures, it is more important to be able to grasp the high-level flow rather than the individual methods/classes.