Add "End Setup" and shortcuts to setups as buttons in the sidebar
The setup buttons are mostly there to help testing things on WASM and iOS since these platforms do not have a window menu bar. This removes the biggest hurdle for testing on these platforms and gives us a bit more time to think about how menu options should be displayed.