Fix problem with ball state being out of sync when picked up
This just fixes the immediate problem, but it does point to a larger problem with the ball location needing to be tracked in multiple places + you need to update both BallState and BallLocation.
The first is due to the fact we do not want to pass a game instance around to all FieldCoordinates, so probably just the less of two evils. But maybe there is a way to combine SetBallState and
SetBallLocation that would make it less likely for bugs to happen.