Refactor skills so each player skill now has unique skillId and rerollSourceId
In order to be able to locate instances of objects across the network protocol, we need everything to have an unique (deterministic) id. A bug was discovered in rerolls because it was sending the entire TeamReroll object. Which then wasn't counted correctly on the receiving side. That bug isn't fixed by this commit, but this is the first step in everything having unique IDs.

This is also the first step towards simplifying what needs to be serialized. Right now, we are serializing Game, Team and Player. And all of these classes have a lot of problems because of this (as circular dependencies needs to be broken). So ideally Game should never be serialized, but only the initial Team data and GameActions (which should be refactored to only contains ID's and never full objects.