Add Network Fuzzer
Adding a way to fuzz test the network traffic should help weed out issues with serialization. There are probably still room for errors in the action providers, but that should be tested in a different test. This refactor also make it possible to define the random generator and pass it in from the top-level.