Refactor event handling in P2P Game menus
Before this commit, event flow inside the P2PHost and P2PClient was a huge mess. At least this cleans it up somewhat so now we control all state changes inside Workflow classes. This makes it easier to reason about what happens when. It still isn't great though and probably needs a new iteration, but first think about the best UI flow. Right now there is a lot of complexity because you can jump between a lot of states. That is probably not needed.

Also, right now the boundary between menus and the actual game is quite messy and going back to the menu from the game has a huge chance to break a lot of things. We should create a more hard boundary so leaving the game screen is counted as a disconnect.