Add new `skipNodeFor` method in ParentNode
This helps encoding a common pattern where you would have a ComputationNode deciding whether or not to go into a ParentNode or skip to the next. Now it can be encoded directly in the ParentNode which reduces the total amount of nodes, reducing the size of the FSM and hopefully makes the rules flow more clear.