Hide Red and Yellow glows (for Knocked Down and Active Menu)
This is mostly testing how it feels, the Knocked Down is partly not needed anymore for POW (since we have a blocked indicator), it might be useful for Player Down, but since that player is always sideways there is already an indication. Same for dodges, GFI etc.

Active Menu is only relevant for active players anyway, so also a bit duplicated.

It is mostly a manner of figuring out how clear it should be which player is "in the center"