Make Player Stats Box and Action Wheel diameter scale with the size of the window.
Especially the Player Stats Box would look very bad when scaled to smaller sizes, and since it doesn't contain any interactive elements, just scaling with the size of the window seems fine.

Scaling the Action wheel is a bit more controversial, but keeping it the same size across a big 4K screen and smaller lap top sizes looked pretty wrong. So doing it this way is just first attempt. Perhaps we need to introduce a 3rd kind of unit that can scale upwards, but has a lower limit.