Move all menu bar items to an in-game drawer. Undo and Menu have been exposed directly on the game screen.
Before this change, only Desktop had a menu bar. This wasn't visible on Web and iOS. So this change moves all these items to an in-game NavigationDrawer. This is probably the way forward, but I suspect the UI has to change when we add more items there.

I also exposed some developer options that will make it easier for users to report bugs and get insight into the current state of the rules engine.