Add new dogout background + rework dugout layout
This also removed the padding between the field an other components. The idea was to have the distance there to be able to make a ball go out of field. This can be still be done, albeit it will be slightly more tricky. But given that the distance wouldn't scale well to different square sizes (e.g BB7). It is probably fine having the ball land at the edge of the dugout.