Improve error reporting from the UI
Instead of just opening GitHub with a prefilled form, this commit now changes it so there is an in-game dialog for creating issues.
All issues are then forwarded to the Jervis website where a Github App called the Jervis Support Bot is responsible for creating the issue on GitHub.
Since GitHub does not allow us to attach files through the REST API, all files are instead stored on the web-server and a link is embedded in the issue.

Once created, the URL is returned to the client that will open a browser with the link to the created issue.

It is a concern that creating issues can be done without authentication, but right now the primary goal is to make it as easy as possible to report issues. If it turns out to be a problem, this functionality can be easily disabled by removing the API endpoint at https://jervis.ilios.dk/create_issue.php.