Fix context menu preventing square overlays when undo'ing actions.
This only happened if the context menu was open, you called Undo and the resulting state didn't show a context menu. In that case the "available squares" color effect didn't show.
There is a bigger problem with how to deal with dialog lifecycles during actions and Undo, but it isn't a high priority, it is mostly to make animations nicer.