Move all BB2020 procedures to common
This is in preparation for adding BB2025 specific skills and rules. Many of them will not change from BB2020, so having the starting point being "common" is just easier.
It will not be correct for some of them, but those will be moved back to BB2020 package when we get to their implementation