Add Player Editor context menu
Available when right-clicking on players. Makes it possible to edit player skills, stats and keywords during a game.
It is hidden behind a Dev setting in Game Customization and only works for hotseat games for now

This also caused a refactor of skill settings. SkillFactories now contain the "generic" name for skills. This is needed to better model skills like "Hatred(X)" in the UI